Dashboard App
Created Dashboard app which uses firebase, user can log in or sing up to the app and submit a post, id user logs in as admin, they can access Dashboard screen where they can approve or reject the message. Almost everything is complete but when I try to change the "Post" object it does not change but makes a new case in firestore.